v1/crypto_bot/requirements.txt
# Crypto Bot — Production Dependencies (Pinned)
ccxt==4.5.38
xgboost==2.1.4
scikit-learn==1.6.1
pandas==2.3.3
numpy==2.0.2
requests==2.32.5
python-dotenv==1.2.1
